feat: Add ability to retrieve all login data and sort it dynamically
This commit adds a new method `get_all_login` to retrieve all the login data and sorts it dynamically. This change has been made to `db_model.php`. Additionally, `index.php` has been modified to use the new method to display login data in the view when a user is logged in.